Renamed example packages.
