+++ examples/jats-editor/package.js
