+++ examples/science-writer/package.js
